requirements.txt
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0rc1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
future==1.0.0
idna==3.7
Jinja2==3.1.4
junos-eznc==2.7.1
lxml==5.2.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
napalm==5.0.0
ncclient==0.6.15
netaddr==1.3.0
netmiko==4.4.0
netutils==1.9.0
nornir==3.4.1
nornir-csv==0.3.0
nornir-netmiko==1.0.1
nornir-utils==0.2.0
nornir_napalm==0.5.0
ntc_templates==6.0.0
numpy==2.0.1
pandas==2.2.2
paramiko==3.4.0
pycparser==2.22
pyeapi==1.0.2
pymongo==4.8.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scp==0.15.0
six==1.16.0
textfsm==1.1.3
tqdm==4.66.4
transitions==0.9.1
ttp==0.9.5
ttp-templates==0.3.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yamlordereddictloader==0.4.2
